+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/product/MaintainProductsPresenter.groovy
@@ -1,6 +1,7 @@
 package life.qbic.portal.offermanager.components.product
 
 import life.qbic.business.products.archive.ArchiveProductOutput
+import life.qbic.business.products.copy.CopyProductOutput
 import life.qbic.business.products.create.CreateProductOutput
 import life.qbic.datamodel.dtos.business.services.Product
 import life.qbic.portal.offermanager.components.AppViewModel
@@ -13,7 +14,7 @@ import life.qbic.portal.offermanager.components.AppViewModel
  * @since 1.0.0
  *
  */
-class MaintainProductsPresenter implements CreateProductOutput, ArchiveProductOutput{
+class MaintainProductsPresenter implements CreateProductOutput, ArchiveProductOutput, CopyProductOutput{
 
     private final MaintainProductsViewModel productsViewModel
     private final AppViewModel mainViewModel
@@ -36,6 +37,12 @@ class MaintainProductsPresenter implements CreateProductOutput, ArchiveProductOu
     }
 
     @Override
+    void copied(Product product) {
+mainViewModel.successNotifications << "Successfully copied product $product.productId - $product.productName."
+        productsViewModel.productsResourcesService.addToResource(product)
+    }
+
+    @Override
     void foundDuplicate(Product product) {
         mainViewModel.failureNotifications << "Found duplicate product for $product.productId - $product.productName."
         //todo triggers sth in the view-model to ask the user if he still wants to create the duplicate
@@ -45,4 +52,5 @@ class MaintainProductsPresenter implements CreateProductOutput, ArchiveProductOu
     void failNotification(String notification) {
         mainViewModel.failureNotifications << notification
     }
+
 }