+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Proteomic and Metabolomic Products can now be selected and included in an Offer (`#425 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/425>`_)
+
 * Link offers to project now. The ``life.qbic.business.offers.Offer`` and ``life.qbic.portal
 .offermanager.dataresources.offers`` have been extended with a new property to associate it with
 an existing project by its project identifier.
@@ -18,6 +20,7 @@ use cases of the product maintenance and creation feature(`#411 <https://github.
 * After a project has been created from an offer, the offer overview is updated accordingly
 (`#427 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/427>`_)
 * Add the UpdatePersonView to separate the Update and Create Person use cases more consequently (`#436 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/436>`_)
+* Proteomic and Metabolomic Products are now included in the Offer PDF
 
 **Fixed**